review: uintptr_t -> unsigned int.

+++ test/extensions/filters/http/wasm/test_data/test_panic_cpp.cc
@@ -33,7 +33,7 @@ static RegisterContextFactory register_PanicContext(CONTEXT_FACTORY(PanicContext
 
 // With Emscripten, dereferencing a null pointer does not immediately cause a segmentation fault,
 // so use an invalid address to trigger it.
-static uintptr_t* badptr = reinterpret_cast<uintptr_t*>(0xDEADBEEF);
+static unsigned int* badptr = reinterpret_cast<unsigned int*>(0xDEADBEEF);
 
 FilterHeadersStatus PanicContext::onRequestHeaders(uint32_t, bool) {
   *badptr = 0;
